+++ engine/tests/skipdata/expression_tree_test.cpp
@@ -1,6 +1,6 @@
 #include <gtest/gtest.h>
-//#include "skip_data/expression_tree.hpp"
 #include "parser/expression_tree.hpp"
+#include "skip_data/SkipDataProcessor.h"
 
 using namespace ral;
 using namespace skip_data;
@@ -13,7 +13,7 @@ struct ExpressionTreeTest : public ::testing::Test {
   }
 
   void SetUp() override {
-  
+
   }
 
   void process(std::string prefix, std::string expected, bool valid_expr = true) {
@@ -21,7 +21,7 @@ struct ExpressionTreeTest : public ::testing::Test {
     tree.build(prefix);
       std::cout << "before:\n";
       tree.print();
-      tree.apply_skip_data_rules();
+      ral::skip_data::apply_skip_data_rules(tree);
       std::cout << "after:\n";
       tree.print();
       auto solution =  tree.prefix();
@@ -75,7 +75,7 @@ TEST_F(ExpressionTreeTest, expr_test_1) {
   std::string prefix = "=(+($0, $1), 123)";
   std::string expected = "AND <= + $0 $2 123 >= + $1 $3 123";
   process(prefix, expected);
-} 
+}
 
 TEST_F(ExpressionTreeTest, expr_test_2) {
   std::string prefix = "OR(AND(AND(>($0, 100), =(+($0, $1), 123)), <($1, 10)), =($0, 500))";
@@ -97,7 +97,7 @@ TEST_F(ExpressionTreeTest, expr_test_4) {
 
 TEST_F(ExpressionTreeTest, expr_test_5) {
   std::string prefix = "OR(AND(AND(>($0, 100), =(*($0, $1), 123)), <($1, 10)), /($0, 500))";
-  std::string expected = "";
+  std::string expected = "NONE";
   process(prefix, expected);
 }
 
@@ -125,27 +125,27 @@ TEST_F(ExpressionTreeTest, expr_test_5) {
 TEST_F(ExpressionTreeTest, expr_test_9) {
   std::string prefix = "OR(>($0, 100), =($0, 500))";
   std::string expected = "OR > $1 100 AND <= $0 500 >= $1 500";
-  bool valid_expr = true;  
+  bool valid_expr = true;
   process(prefix, expected, valid_expr);
 }
 
 TEST_F(ExpressionTreeTest, drop_test1) {
   std::string prefix = "OR(AND(AND(>($0, 100), =(+($0, $1), 123)), <($1, 10)), =($0, 500))";
   std::string expected = "OR > $1 100 AND <= $0 500 >= $1 500";
-  bool valid_expr = true; 
+  bool valid_expr = true;
   ral::parser::parse_tree tree;
   tree.build(prefix);
     std::cout << "before:\n";
     tree.print();
-    tree.drop({"$1"});
+    ral::skip_data::drop_value(tree, "$1");
     std::cout << "after drop $1:\n";
 
     tree.print();
-    
+
     std::string solution =  tree.prefix();
     std::cout << "after drop solution:\n";
     std::cout << solution << "\n";
-    tree.apply_skip_data_rules();
+    ral::skip_data::apply_skip_data_rules(tree);
     std::cout << "after skip_data rules:\n";
     tree.print();
     solution =  tree.prefix();
@@ -158,21 +158,21 @@ TEST_F(ExpressionTreeTest, drop_test1) {
 TEST_F(ExpressionTreeTest, drop_test2) {
   std::string prefix = "OR(AND(AND(>($2, 100), =(+($0, $1), 123)), <($1, 10)), =($0, 500))";
   std::string expected = "OR > $5 100 AND <= $0 500 >= $1 500";
-  bool valid_expr = true; 
+  bool valid_expr = true;
   ral::parser::parse_tree tree;
 
   tree.build(prefix);
     std::cout << "before:\n";
     tree.print();
-    tree.drop({"$1"});
+    ral::skip_data::drop_value(tree, "$1");
     std::cout << "after drop $1:\n";
 
     tree.print();
-    
+
     std::string solution =  tree.prefix();
     std::cout << "after drop solution:\n";
     std::cout << solution << "\n";
-    tree.apply_skip_data_rules();
+    ral::skip_data::apply_skip_data_rules(tree);
     std::cout << "after skip_data rules:\n";
     tree.print();
     solution =  tree.prefix();
@@ -183,20 +183,21 @@ TEST_F(ExpressionTreeTest, drop_test2) {
 
 TEST_F(ExpressionTreeTest, drop_test3) {
   std::string prefix = "AND(=(+($0, $1), 123), =($0, $1))";
-  std::string expected = "";
+  std::string expected = "NONE";
    ral::parser::parse_tree tree;
    tree.build(prefix);
     std::cout << "before:\n";
     tree.print();
-    tree.drop({"$0", "$1"});
+    ral::skip_data::drop_value(tree, "$0");
+    ral::skip_data::drop_value(tree, "$1");
     std::cout << "after drop $1:\n";
 
     tree.print();
-    
+
     std::string solution =  tree.prefix();
     std::cout << "after drop solution:\n";
     std::cout << solution << "\n";
-    tree.apply_skip_data_rules();
+    ral::skip_data::apply_skip_data_rules(tree);
     std::cout << "after skip_data rules:\n";
     tree.print();
     solution =  tree.prefix();
